Finished Kotlin part FFmpegDecoder.

+++ src/main/kotlin/charlie/laplacian/decoder/Decoder.kt
@@ -8,8 +8,8 @@ interface Decoder {
     fun positionMillis(): Long
     fun durationMillis(): Long
 
-    fun volumeTo(percent: Long)
-    fun volume(): Long
+    fun volumeTo(percent: Int)
+    fun volume(): Int
 
     fun close()
 }
